pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.2.0"
description = "A simple ToDoList application with database storage"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
click = "^8.1.0"
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"
schedule = "^1.2.0"
psycopg2-binary = "^2.9.11"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.alembic]
script_location = "alembic"
prepend_sys_path = "."
version_path_separator = "os"
